refactor: move WebhookRequestData to models, cleanup unused imports
